Move TopBar/Footer parts back to src/pages/workspaces.
